% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_explorer.R
\name{model_explorer}
\alias{model_explorer}
\title{Model Explorer}
\usage{
model_explorer(
  data,
  y_var,
  x1_var,
  x2_var,
  x3_var = NULL,
  x4_var = NULL,
  x5_var = NULL,
  p_value_threshold = 0.05
)
}
\arguments{
\item{data}{A data frame containing the variables to be used in the models.}

\item{y_var}{A string indicating the name of the dependent (outcome) variable.}

\item{x1_var}{A string indicating the name of the first independent variable.}

\item{x2_var}{A string indicating the name of the second independent variable.}

\item{x3_var}{An optional string indicating the name of the third independent variable.}

\item{x4_var}{An optional string indicating the name of the fourth independent variable.}

\item{x5_var}{An optional string indicating the name of the fifth independent variable.}

\item{p_value_threshold}{A numeric value specifying the significance threshold for p-values.
Defaults to 0.05.}
}
\value{
A data frame containing significant models:
  \itemize{
  \item{model:} The model formula as a string.
  \item{significant_predictors:} Character vector of significant predictor names.
  \item{p_values:} Numeric vector of p-values for significant predictors.
  }
  If no significant predictors are found, a message is printed.
}
\description{
This function systematically explores various linear models using a given dataset
to reveal significant predictors and interactions.
}
\examples{
# Assuming sample data called 'my_data' with variables
#  'outcome', 'pred1', 'pred2'
model_explorer(data = my_data, y_var = "outcome", x1_var = "pred1", x2_var = "pred2")

}
