% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shadow_read.R
\name{shadow_read}
\alias{shadow_read}
\title{Analyze Sensitivity Analysis Output}
\usage{
shadow_read(results, top_n = NULL)
}
\arguments{
\item{results}{Output list from a sensitivity analysis function.}

\item{top_n}{Optional. An integer controlling the number of top excluded/included observations to report. Defaults to 10% of the observations in the original dataset.}
}
\value{
A list containing:
\itemize{
  \item excluded_frequencies: A list of exclusion frequency vectors for each shadow dataset.
  \item included_frequencies: A list of inclusion frequency vectors for each shadow dataset.
  \item total_excluded: Total exclusion frequencies across all datasets.
  \item total_included: Total inclusion frequencies across all datasets.
  \item top_excluded_obs: IDs of the top excluded observations.
  \item top_included_obs: IDs of the top included observations.
  \item top_excluded_df: Data frame containing rows for the top excluded observations.
  \item top_included_df: Data frame containing rows for the top included observations.
}
}
\description{
Calculates observation inclusion/exclusion frequencies from the results of a sensitivity analysis from the `ten_shadows` function.
}
