% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ten_shadows.R
\name{ten_shadows}
\alias{ten_shadows}
\title{Sensitivity Analysis with Randomized Subsets}
\usage{
ten_shadows(
  dataset,
  predictors,
  outcome,
  controls = NULL,
  interactions = NULL,
  num_obs_remove = NULL,
  num_iterations = 10,
  p_value_threshold = 0.05,
  check_interaction = FALSE
)
}
\arguments{
\item{dataset}{A data frame containing the variables for the sensitivity analysis.}

\item{predictors}{A character vector specifying the predictor variables of interest.}

\item{outcome}{A character string specifying the outcome (dependent) variable.}

\item{controls}{A character vector (optional) specifying control variables to include in the model.}

\item{interactions}{A character vector (optional) of variable interactions. Interaction
terms should be formatted as `"var1:var2"`. Ensure main effects for interactions are
included within `predictors` or `controls`.}

\item{num_obs_remove}{An integer (optional) specifying the number of observations to remove
in each iteration. If not provided, 10% of observations will be removed.}

\item{num_iterations}{An integer specifying the number of model iterations to perform.
Defaults to 10.}

\item{p_value_threshold}{A numeric value indicating the significance threshold for
determining a "significant" shadow dataset. Defaults to 0.05.}

\item{check_interaction}{A logical value. If `TRUE`, checks for significant interactions
(if interaction terms are provided). Defaults to `FALSE`.}
}
\value{
A list containing all shadow datasets where either predictors or their specified
  interactions (if applicable) were found to be statistically significant. Each list
  element is named "shadow_1", "shadow_2", etc. and contains:
  * **dataset:** The subset of data used for that iteration.
  * **model_summary:** The output from `summary(lm())`.
}
\description{
The `ten_shadows` function investigates the robustness of statistical relationships by
repeatedly fitting a linear model to randomly reduced subsets of the original dataset.
This method helps identify cases where findings may be sensitive to the inclusion or
exclusion of specific observations.
}
\details{
Each iteration involves randomly removing observations, refitting the
  linear model, and assessing significance. If no shadow datasets reach significance,
  an error is returned indicating potential robustness of the predictors.
}
\examples{
\dontrun{
# Example usage with the 'mtcars' dataset
results <- ten_shadows(dataset = mtcars,
                       predictors = c("mpg", "hp"),
                       outcome = "wt",
                       controls = "cyl")
}

}
